agent_waha/requirements.txt
flask
python-dotenv
openai==1.58.1
aiohttp
requests
langchain
datetime
tinydb
pydantic
langchain-community
langchainhub
langgraph
langchain-openai
langchain-core
typing
pymongo
langchain-chroma
decouple
unstructured
docx2txt
python-docx
langgraph-checkpoint-sqlite
langgraph-checkpoint-mongodb==0.1.3
langchain-mongodb
langchain-experimental
datetime
google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib
plotly
pandas
openpyxl
setuptools
boto3
ipykernel
unidecode
fake-useragent
tabulate